refactor: 諸々 utils に移動
